pyproject.toml
[project]
name = "fracnum"
description = "Tools for fractional calculus and fractional differential equations. By Niels Goedegebure."
version = "0.9.5"

requires-python = ">=3.8"

dependencies = [
    "numpy==1.26.4",
    "scipy~=1.13.1",
    "matplotlib~=3.9.2",
    "mpmath~=1.3.0",
    "tqdm~=4.66.5",
    "xarray~=2024.7.0",
    "netcdf4~=1.7.2",
    "python-dotenv==1.1.0"
]

[tool.setuptools]
packages = ["fracnum"]